move postgres repositories test case into tests package in postgres directory and clean some code
